Add the internal functionality to handle tables and description lists
This does not add description lists and tables to the base functionality, but it provides the hooks to implement it and some examples with it partially implemented.

The key features is the ability to override the functionality of the enter and tab key when the selection is within certain blocks. And the added ability to treat \t (tab character) as a new sub-line type which better mimics the CSV format used for tabular data.